+++ environ/environ.h
@@ -6,7 +6,7 @@
 /*   By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/09 15:38:45 by mhirabay          #+#    #+#             */
-/*   Updated: 2022/02/10 14:15:56 by mhirabay         ###   ########.fr       */
+/*   Updated: 2022/02/10 15:37:40 by mhirabay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,6 +44,8 @@ char		*get_value(void *content);
 char		*get_key(void *content);
 bool		is_lvalue_bigger_ascii(char *lvalue, char *rvalue);
 bool		is_same_key(char *a, char *b);
+
+// kvs_utils_2.c
 void		*create_kvs_content(char *key, char *value);
 
 // lst_utils.c
