+++ environ/kvs_utils.c
@@ -6,7 +6,7 @@
 /*   By: mhirabay <mhirabay@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/04 14:43:17 by mhirabay          #+#    #+#             */
-/*   Updated: 2022/02/10 13:04:01 by mhirabay         ###   ########.fr       */
+/*   Updated: 2022/02/10 15:36:43 by mhirabay         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,31 +64,4 @@ bool	is_same_key(char *a, char *b)
 	if (ft_strncmp(a, b, ft_strlen(a)) != 0)
 		return (false);
 	return (true);
-}
-
-void	*create_kvs_content(char *key, char *value)
-{
-	t_kvs	*kvs;
-
-	kvs = (t_kvs *)malloc(sizeof(t_kvs));
-	if (kvs == NULL)
-		return (NULL);
-	kvs->key = ft_strdup(key);
-	if (kvs->key == NULL)
-		return (NULL);
-	// export k のような形で入る場合を考慮して、valueはnullを許容する
-	if (value != NULL)
-	{
-		kvs->value = ft_strdup(value);
-		if (kvs->value == NULL)
-		{
-			free(kvs->key);
-			return (NULL);
-		}
-	}
-	else
-	{
-		kvs->value = NULL;
-	}
-	return ((void *)kvs);
 }
